refactor: avoid rebuilding when build info changed.
